+++ src/main/java/step3/view/ResultView.java
@@ -1,8 +1,11 @@
-package step3.car.game;
+package step3.view;
+
+import step3.domain.Car;
+import step3.domain.Cars;
 
 import java.util.stream.IntStream;
 
-public class CarsView {
+public class ResultView {
 
     public void print(Cars cars) {
         cars.cars()
