+++ src/test/java/step3/car/game/unit/CarTest.java
@@ -3,8 +3,8 @@ package step3.car.game.unit;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.car.game.Car;
-import step3.car.game.Cars;
+import step3.domain.Car;
+import step3.domain.Cars;
 
 import static org.assertj.core.api.Assertions.*;
